+++ src/main/extensions/blow/command/ClusterCommands.groovy
@@ -22,7 +22,6 @@ package blow.command
 import blow.BlowSession
 import blow.shell.BlowShell
 import blow.shell.Cmd
-import blow.shell.Synopsis
 import groovy.util.logging.Slf4j
 
 /**
@@ -37,8 +36,7 @@ class ClusterCommands {
     def BlowShell shell
     def BlowSession session
 
-	@Cmd
-    @Synopsis("Create and launch the cluster using the settings provided in the configuration file")
+	@Cmd( summary = "Create and launch the cluster using the settings provided in the configuration file" )
 	public void start() {
         def size = session.conf.size
         def answer = shell.prompt("Please confirm that you want to start ${size} node(s) [y/n]") { ['y','n'].contains(it) }
@@ -50,8 +48,7 @@ class ClusterCommands {
 
 
 
-    @Cmd
-    @Synopsis("Terminate the current cluster")
+    @Cmd(summary="Terminate the current cluster")
     public void terminate() {
 
         def answer = shell.prompt("Please confirm that you want to termiante cluster named ${shell.session.clusterName} [y/n]") {
@@ -63,8 +60,7 @@ class ClusterCommands {
         }
     }
 
-    @Cmd
-    @Synopsis("Shows the list of the current running clusters")
+    @Cmd(summary="Shows the list of the current running clusters")
     def listclusters() {
 
         def clusters = session.listClusters()
@@ -79,8 +75,7 @@ class ClusterCommands {
 
 
 
-    @Cmd
-    @Synopsis("Print the configuration settings for the current selected cluster")
+    @Cmd(summary="Print the configuration settings for the current selected cluster")
     def conf() {
         println session.getConfString()
     }