+++ src/main/extensions/blow/command/StorageCommands.groovy
@@ -19,15 +19,17 @@
 
 package blow.command
 
-import org.jclouds.ec2.domain.Volume
-import blow.shell.Cmd
+import blow.BlowSession
 import blow.shell.BlowShell
-import org.jclouds.ec2.domain.Snapshot
+import blow.shell.Cmd
 import blow.shell.Completion
 import blow.shell.Opt
-import blow.shell.Synopsis
+import org.jclouds.ec2.domain.Attachment
+import org.jclouds.ec2.domain.Snapshot
+import org.jclouds.ec2.domain.Volume
 
 /**
+ * Shell command managing EBS volumes and snapshots
  *
  * @author Paolo Di Tommaso
  */
@@ -35,25 +37,67 @@ class StorageCommands  {
 
     // injected by the framework
     def BlowShell shell
+    def BlowSession session
 
     /**
+     * Prints the list of volumes
      *
-     * @param args
+     * @param regionId
+     *      The region for which list the volumes, if null will be listed the volumes
+     *      for the current configured region
+     * @param attachments
+     *      If true it will print the attachment (if any) for the volume
+     * @param zone
+     *      The zone in which the volumes is allocated
+     * @param volumeIds
+     *      Shows only the volumes which matches the specified ids
      */
-    @Cmd("listvolumes")
+    @Cmd( name="listvolumes",
+          summary="Display the list of volumes in the current region",
+          usage="listvolumes [options] [volume-id ..]")
     @Completion( { findVolumesCompletion(it) } )
-    @Synopsis("Display the list of volumes in the current region")
-    def void listVolumes( @Opt( name='name') def name,  def args ) {
 
-        def list = shell.session.blockStore.listVolumes()
+    def void listVolumes(
+            @Opt(opt="r", longOpt="region", arg="region-id", description="Specify a different region")
+            String regionId,
+            @Opt(opt='a',longOpt='attachment', description='Include the attachments in thge report')
+            Boolean attachments,
+            @Opt(opt='z', longOpt='zone', description='Include the availability zone in the report')
+            Boolean zone,
+            List<String> volumeIds )
+    {
+
+        def list = shell.session.blockStore.listVolumes(volumeIds, regionId)
+
         if( !list ) {
             println "(no volumes found)"
-
+            return
         }
 
         list.each() { Volume vol ->
-            def size = "${vol.size} G".padLeft(5)
-            println "${vol.id}; ${size}; ${vol.status}"
+            def line = new StringBuilder()
+            line.append(vol.id).append("; ")
+            line.append("${vol.size} G".padLeft(5)) .append("; ")
+            line.append(vol.getCreateTime()?.format('yyyy-MM-dd HH:mm')) .append("; ")
+            line.append(vol.getStatus()?.toString().padLeft(9) ) .append("; ")
+
+            // print out the availability zone
+            if( zone ) {
+                line.append(vol.getAvailabilityZone()) .append('; ')
+            }
+
+            // print out the attachments if required
+            if( attachments ) {
+                vol.getAttachments().each { Attachment att ->
+
+                    line.append("\n  > attached to: ")
+                        .append( att.id ).append(' - ')
+                        .append( att.device )
+                        .append(' (').append( att.status ).append(')')
+                }
+            }
+
+            println line
         }
     }
 
@@ -75,19 +119,148 @@ class StorageCommands  {
         vols.collect { vol -> vol.id }
     } 
 
+    /**
+     * Prints the list of the snapshots
+     *
+     * @param printDescription
+     * @param printVolume
+     * @param snapshotIds
+     */
 
-    @Cmd("listsnapshots")
-    @Synopsis("Display the list of snapshots available in the current region")
-    def void listSnapshots() {
+    @Cmd( name="listsnapshots",
+          summary="Display the list of snapshots available in the current region",
+          usage="listsnapshots [options]")
+    
+    def void listSnapshots(
 
-        def list = shell.session.blockStore.listSnapshots()
+            @Opt(opt="d", longOpt="description", description="Include the snapshot 'description' in the report")
+            Boolean printDescription,
+
+            @Opt(opt="v", longOpt="volume", description="Include the associated volume-id in the report")
+            Boolean printVolume,
+            List<String> snapshotIds )
+    {
+
+        def list = shell.session.blockStore.listSnapshots(snapshotIds)
         if( !list ) {
             println "(no volumes found)"
-
+            return
         }
 
+        /*
+         * print a row for each snapshots
+         */
         list.each() { Snapshot snapshot ->
-            println "${snapshot.id} [${snapshot.volumeId}] ${snapshot.status}"
+            def line = new StringBuilder(snapshot.id)
+                    .append("; ")
+                    .append(snapshot.getStartTime()?.format('yyyy-MM-dd HH:mm')) .append("; ")
+                    .append(snapshot.getStatus()) .append("; ")
+
+            /*
+             * include the volume info if required
+             */
+            if( printVolume ) {
+                def vol = snapshot.getVolumeId() ?: ""
+                line.append(vol.padRight(8)).append("; ")
+            }
+
+            /*
+             * include the snapshot description if required
+             */
+            if( printDescription ) {
+                def txt = snapshot.getDescription()
+                line.append( txt ? "\"$txt\"" : "")
+            }
+
+            println line
         }
     }
+
+    /**
+     * Create a new snapshot
+     *
+     * @param description
+     *      A string value representing the snapshot description
+     * @param volumeId
+     *      The volume-id from which create the snapshot
+     */
+    @Cmd( name="createsnapshot",
+          summary="Create a snapshot from given the specified EBS volume id",
+          usage="createsnapshot [options] snapshot-id")
+
+    @Completion( { findVolumesCompletion(it) } )
+
+    def void createSnapshot(
+            @Opt(opt='d', longOpt='description', len=1, description="Snapshot description (shorter than 255 chars)")
+            String description,
+            String volumeId
+    ) {
+        final ebs = shell.session.blockStore
+
+        if( !volumeId ) {
+            println "You need to specify the volume-id as command parameter"
+            return
+        }
+
+        def vol = ebs.findVolume(volumeId)
+        if( !vol ) {
+            println "Cannot find any volume with id: '$volumeId' in region: '${session.conf.regionId}'"
+            return
+        }
+
+        if( description?.length() > 255 ) {
+            println "Please provide a shorter description"
+            return
+        }
+
+        def answer = shell.promptYesOrNo("You are going to create a snapshot for the volume: ${volumeId}. Do you want to continue?")
+
+        if( answer != 'y' ) { return }
+
+        /*
+         * create the snapshot
+         */
+
+        def snapshot = ebs.createSnapshot(volumeId, description, false)
+
+        println "Creating snapshot: ${snapshot.getId()}. It can take some minutes to complete."
+
+    }
+
+    /**
+     * Delete a snapshot store from the AWS storage
+     *
+     * @param snapshotId
+     *      The id of the snapshot to delete
+     */
+    @Cmd( name='deletesnapshot',
+          summary = 'Delete a AWS snapshot',
+          usage='deletesnapshot snapshot-id')
+    def void deletesnapshot( String snapshotId ) {
+        if( !snapshotId ) {
+            println "Provide on the command line the id of the snapshot to delete."
+            return
+        }
+
+
+        def answer = shell.prompt("You are going to DELETE the snapshot '${snapshotId}'. Please enter the snapshot-id to confirm:", [snapshotId,'n'])
+        if( answer != snapshotId ) {
+            return
+        }
+
+        Snapshot snap = session.blockStore.deleteSnapshot(snapshotId);
+        if( snap == null ) {
+            print "Cannot delete snapshot '${snapshotId}'. See the log file for details."
+            return
+        }
+
+        println "Snapshot schedueled for deletion. It can take some minutes."
+    }
+
+
+//    dev void deleteVolume( String volumeId ) {
+//
+//        session.blockStore.deleteVolume()
+//    }
+
 }